create the triangle subsystem
because:
- this is generally the pattern I'll try to follow
  - the frame is responsible for all frame-specific resources,
    things like buffer pools and command buffers
  - a RenderTarget is a thing which knows how to use a frame to draw
    something interesting.
